refactor: extract main tab policies
