+++ apps/ios/LogYourBody/Views/MainTabView.swift
@@ -5,18 +5,6 @@
 import SwiftUI
 import CoreData
 
-enum PaidAppSurface: Equatable {
-    case weightLoggerMVP
-    case legacyFullDashboardBeta
-    case photoTimelineHUD
-}
-
-enum PaidAppSurfacePolicy {
-    static func surface() -> PaidAppSurface {
-        .photoTimelineHUD
-    }
-}
-
 struct MainTabView: View {
     @AppStorage("healthKitSyncEnabled") private var healthKitSyncEnabled = true
     @EnvironmentObject var authManager: AuthManager
@@ -64,45 +52,6 @@ struct MainTabView: View {
         selectedSurface = PaidAppSurfacePolicy.surface()
 
         HealthSyncCoordinator.shared.bootstrapIfNeeded(syncEnabled: healthKitSyncEnabled)
-    }
-}
-
-enum PaidWeightLoggerMVPPolicy {
-    static func validationMessage(weightText: String, unit: String) -> String? {
-        let trimmed = weightText.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty else { return nil }
-
-        do {
-            _ = try ValidationService.shared.validateWeight(trimmed, unit: unit)
-            return nil
-        } catch {
-            return error.localizedDescription
-        }
-    }
-
-    static func canSaveWeight(weightText: String, unit: String, isSaving: Bool) -> Bool {
-        let trimmed = weightText.trimmingCharacters(in: .whitespacesAndNewlines)
-        guard !trimmed.isEmpty, !isSaving else { return false }
-        return validationMessage(weightText: trimmed, unit: unit) == nil
-    }
-
-    static func syncStatusText(status: RealtimeSyncManager.SyncStatus, pendingCount: Int) -> String {
-        switch status {
-        case .syncing:
-            return "Syncing"
-        case .success:
-            return "Synced"
-        case .error:
-            return "Sync needs retry"
-        case .offline:
-            return pendingCount > 0 ? "Saved offline" : "Offline"
-        case .idle:
-            return pendingCount > 0 ? "Pending sync" : "Synced"
-        }
-    }
-
-    static func savedConfirmationText(isOnline: Bool) -> String {
-        isOnline ? "Saved locally. Pending sync." : "Saved locally. Will sync when online."
     }
 }
